move search list to shared ui
