+++ cmd/ui/src/views/Explore/ExploreSearch/CommonSearches.tsx
@@ -14,28 +14,14 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    List,
-    ListSubheader,
-    ListItem,
-    ListItemText,
-    ListItemButton,
-    Box,
-    Tabs,
-    Tab,
-    Typography,
-    IconButton,
-} from '@mui/material';
-import { FC, useState } from 'react';
-import { apiClient, CommonSearches as prebuiltSearchList } from 'bh-shared-ui';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faTrash } from '@fortawesome/free-solid-svg-icons';
+import { Box, Tabs, Tab, Typography } from '@mui/material';
+import { useState } from 'react';
+import { PrebuiltSearchList, CommonSearches as prebuiltSearchList } from 'bh-shared-ui';
 import makeStyles from '@mui/styles/makeStyles';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useDispatch } from 'react-redux';
-import { addSnackbar } from 'src/ducks/global/actions';
 import { setCypherQueryTerm } from 'src/ducks/searchbar/actions';
 import { startCypherQuery } from 'src/ducks/explore/actions';
+import PersonalSearchList from './PersonalSearchList';
 
 const AD_TAB = 'Active Directory';
 const AZ_TAB = 'Azure';
@@ -104,126 +90,10 @@ const CommonSearches = () => {
                 <Tab label={CUSTOM_TAB} key={CUSTOM_TAB} value={CUSTOM_TAB} className={classes.tab} />
             </Tabs>
 
-            {activeTab === AD_TAB && <SearchList listSections={adSections} onClickListItem={handleClick} />}
-            {activeTab === AZ_TAB && <SearchList listSections={azSections} onClickListItem={handleClick} />}
-            {activeTab === CUSTOM_TAB && <PersonalSearchList onClickListItem={handleClick} />}
+            {activeTab === AD_TAB && <PrebuiltSearchList listSections={adSections} clickHandler={handleClick} />}
+            {activeTab === AZ_TAB && <PrebuiltSearchList listSections={azSections} clickHandler={handleClick} />}
+            {activeTab === CUSTOM_TAB && <PersonalSearchList clickHandler={handleClick} />}
         </Box>
-    );
-};
-
-interface SearchListProps {
-    listSections: ListSection[];
-    onClickListItem: (query: string) => void;
-
-    deleteHandler?: any;
-}
-
-type ListSection = {
-    subheader: string;
-    lineItems: LineItem[];
-};
-
-type LineItem = {
-    id?: number;
-
-    description: string;
-    cypher: string;
-    canEdit?: boolean;
-};
-
-const SearchList: FC<SearchListProps> = ({ listSections, onClickListItem, deleteHandler }) => {
-    const classes = useStyles();
-
-    return (
-        <List dense disablePadding className={classes.list}>
-            {listSections.map((section) => {
-                const { subheader, lineItems } = section;
-
-                return (
-                    <Box key={subheader}>
-                        <ListSubheader sx={{ fontWeight: 'bold' }}>{subheader} </ListSubheader>
-
-                        {lineItems?.map((lineItem, idx) => {
-                            const { id, description, cypher, canEdit = false } = lineItem;
-
-                            return (
-                                <ListItem
-                                    disablePadding
-                                    key={`${id}-${idx}`}
-                                    secondaryAction={
-                                        canEdit && (
-                                            <IconButton size='small' onClick={() => deleteHandler(id)}>
-                                                <FontAwesomeIcon icon={faTrash} />
-                                            </IconButton>
-                                        )
-                                    }>
-                                    <ListItemButton onClick={() => onClickListItem(cypher)}>
-                                        <ListItemText primary={description} />
-                                    </ListItemButton>
-                                </ListItem>
-                            );
-                        })}
-                    </Box>
-                );
-            })}
-        </List>
-    );
-};
-
-// `PersonalSearchList` is a more specific implementation of `SearchList`.  It includes
-// additional fetching logic to fetch and delete queries saved by the user
-const PersonalSearchList: FC<{ onClickListItem: (query: string) => void }> = ({ onClickListItem }) => {
-    const dispatch = useDispatch();
-    const queryClient = useQueryClient();
-
-    const [queries, setQueries] = useState<LineItem[]>([]);
-
-    useQuery({
-        queryKey: 'userSavedQueries',
-        queryFn: () => {
-            return apiClient
-                .getUserSavedQueries()
-                .then((response) => {
-                    const queries = response.data.data;
-
-                    const queriesToDisplay = queries.map((query) => ({
-                        description: query.name,
-                        cypher: query.query,
-                        canEdit: true,
-                        id: query.id,
-                    }));
-
-                    setQueries(queriesToDisplay);
-                })
-                .catch(() => {
-                    setQueries([]);
-                });
-        },
-    });
-
-    const mutation = useMutation({
-        mutationFn: (queryId: number) => {
-            return apiClient.deleteUserQuery(queryId);
-        },
-        onSettled: () => {
-            queryClient.invalidateQueries({ queryKey: 'userSavedQueries' });
-        },
-        onSuccess: () => {
-            dispatch(addSnackbar(`Query deleted.`, 'userDeleteQuery'));
-        },
-    });
-
-    return (
-        <SearchList
-            listSections={[
-                {
-                    subheader: 'User Saved Searches: ',
-                    lineItems: queries,
-                },
-            ]}
-            onClickListItem={onClickListItem}
-            deleteHandler={mutation.mutate}
-        />
     );
 };
